Standalone minimal version of `trv`.
`trv` is capable and should be responsible for bootstrapping
ALL Afiniate projects. It does not depend on any other Afiniate repos.

After this PR, another one will follow removing from `vrt` those
parts that have moved to `trv`.

This removes the chicken and the egg problem for starting developing
for Afiniate.